// jest-puppeteer-ui-test/lib/search-page/interfaces/web-search-page-config.ts
export interface WebSearchPageConfig {
  lang: string;
  fontRatio: number;
  scene: number;
  version: number;
  qqFaceFolderPath: string;
  platform: string;
  netType: string;
  type: number;
  isHomePage?: 0 | 1;
  query?: string;
  isSug?: boolean;
  isLocalSug?: boolean;
  sceneActionType?: number;
  musicSnsId?: string;
  userName?: string;
  searchId?: string;
  sugId?: string;
  sugClickType?: number;
  isMostSearchBiz?: 1;
  sessionId: string;
  subSessionId: string;
  subScene?: number;
  systemVersion: number;
  wechatVersion: number;
  deviceName: string;
  deviceModel: string;
  imei: string;
  deviceBrand: string;
  ostype: string;
  isOverseaApp: 0 | 1;
  // unknown
  educationTab?: any;
  extParams?: any;
  inputMarginTop?: any;
  inputMarginLeftRight?: any;
  inputHeight?: any;
  nativeHeight?: any;
  recommendId?: string;
  prefixSug?: string;
  isPreload?: any;
  pRequestId?: string;
  redPointMsgId?: string;
  thirdExtParam?: any;
  tabInfo?: any;
  extInfo?: any;
  profile?: any;
  passSvrParamStr?: string;
  reqKey?: any;
  chatSearch?: any;
  cacheImageSrc?: any;
  cursorIndex?: any;
  fromTagSearch?: any;
  extReqParam?: any;
  isFTSH5WebSearch?: any;
  entranceExptValue?: any;
  category?: any;
  talker?: any;
  IsFirstLoadTab?: any;
  numberReddot?: number;
  contentReddot?: any;
  transparentMode?: any;
  extReqParams?: any;
  showLocalHistory?: any;
  isIndependentHomepage?: any;
  isV2?: any;
  hasSearchBar?: 1;
  sugIconLeftPadding?: number;
  isClientLoading?: 1;
}